+++ apps/AndroidApp/app/src/expo/java/com/callstack/brownfield/android/example/ReactNativeConstants.kt
@@ -1,5 +1,0 @@
-package com.callstack.brownfield.android.example
-
-object ReactNativeConstants {
-    const val MAIN_MODULE_NAME = "RNApp"
-}