+++ apps/AndroidApp/app/src/expo/java/com/callstack/brownfield/android/example/ReactNativeHostManager.kt
@@ -1,3 +1,0 @@
-package com.callstack.brownfield.android.example
-
-typealias ReactNativeHostManager = com.callstack.rnbrownfield.demo.expoapp.ReactNativeHostManager