+++ packages/brownfield-navigation/android/src/main/java/com/callstack/nativebrownfieldnavigation/BrownfieldNavigationManager.kt
@@ -1,0 +1,14 @@
+package com.callstack.nativebrownfieldnavigation
+
+object BrownfieldNavigationManager {
+  private var navigationDelegate: BrownfieldNavigationDelegate? = null
+
+  fun setDelegate(navigationDelegate: BrownfieldNavigationDelegate) {
+    this.navigationDelegate = navigationDelegate
+  }
+
+  fun getDelegate(): BrownfieldNavigationDelegate {
+    return navigationDelegate
+      ?: throw IllegalStateException("BrownfieldNavigation delegate is not set.")
+  }
+}