+++ packages/brownfield-navigation/src/NativeBrownfieldNavigation.ts
@@ -1,0 +1,9 @@
+import { TurboModuleRegistry, type TurboModule } from 'react-native';
+
+export interface Spec extends TurboModule {
+  temporary(): void;
+}
+
+export default TurboModuleRegistry.getEnforcing<Spec>(
+  'NativeBrownfieldNavigation'
+);