+++ packages/react-native-brownfield/src/expo-config-plugin/android/constants.ts
@@ -1,2 +1,0 @@
-export const BROWNFIELD_PLUGIN_VERSION = '1.0.0';
-export const brownfieldGradlePluginDependency = `classpath("com.callstack.react:brownfield-gradle-plugin:${BROWNFIELD_PLUGIN_VERSION}")`;