+++ packages/react-native-brownfield/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @callstack/react-native-brownfield
 
+## 3.3.0
+
+### Minor Changes
+
+- [#248](https://github.com/callstack/react-native-brownfield/pull/248) [`b0a6f41`](https://github.com/callstack/react-native-brownfield/commit/b0a6f4185aad4e8759b90ccaf9867be493a979ec) Thanks [@artus9033](https://github.com/artus9033)! - feat: support Expo 55
+
+### Patch Changes
+
+- Updated dependencies [[`b0a6f41`](https://github.com/callstack/react-native-brownfield/commit/b0a6f4185aad4e8759b90ccaf9867be493a979ec)]:
+  - @callstack/brownfield-cli@3.3.0
+
+## 3.2.1
+
+### Patch Changes
+
+- [#265](https://github.com/callstack/react-native-brownfield/pull/265) [`7e114e6`](https://github.com/callstack/react-native-brownfield/commit/7e114e676fc290a54912b4d99bd22952b2f7380d) Thanks [@hurali97](https://github.com/hurali97)! - version bump
+
+- Updated dependencies [[`7e114e6`](https://github.com/callstack/react-native-brownfield/commit/7e114e676fc290a54912b4d99bd22952b2f7380d)]:
+  - @callstack/brownfield-cli@3.2.1
+
+## 3.2.0
+
+### Minor Changes
+
+- [#263](https://github.com/callstack/react-native-brownfield/pull/263) [`b7dfa7e`](https://github.com/callstack/react-native-brownfield/commit/b7dfa7e30e1a921bb7ac0952e23481ee081fb9ed) Thanks [@hurali97](https://github.com/hurali97)! - version bump for release
+
+### Patch Changes
+
+- Updated dependencies [[`b7dfa7e`](https://github.com/callstack/react-native-brownfield/commit/b7dfa7e30e1a921bb7ac0952e23481ee081fb9ed)]:
+  - @callstack/brownfield-cli@3.2.0
+
+## 3.1.0
+
+### Minor Changes
+
+- [#252](https://github.com/callstack/react-native-brownfield/pull/252) [`d5c8609`](https://github.com/callstack/react-native-brownfield/commit/d5c8609485b072eb73be0158d49908fa399d8950) Thanks [@artus9033](https://github.com/artus9033)! - feat: added postMessage API
+
+- [#236](https://github.com/callstack/react-native-brownfield/pull/236) [`3e33b89`](https://github.com/callstack/react-native-brownfield/commit/3e33b890e6f647f6bbebe32f8068a38b1ed85ea0) Thanks [@hurali97](https://github.com/hurali97)! - add brownfield navigation
+
+### Patch Changes
+
+- Updated dependencies [[`3e33b89`](https://github.com/callstack/react-native-brownfield/commit/3e33b890e6f647f6bbebe32f8068a38b1ed85ea0)]:
+  - @callstack/brownfield-cli@3.1.0
+
+## 3.0.0
+
+### Patch Changes
+
+- [#255](https://github.com/callstack/react-native-brownfield/pull/255) [`48358b2`](https://github.com/callstack/react-native-brownfield/commit/48358b2dcce578aa5052e66cc3454524da8c7992) Thanks [@artus9033](https://github.com/artus9033)! - chore: release stable v3
+
+- [#239](https://github.com/callstack/react-native-brownfield/pull/239) [`828bda3`](https://github.com/callstack/react-native-brownfield/commit/828bda32f39153085f5053becc7e9de16d284be5) Thanks [@hurali97](https://github.com/hurali97)! - fix: autodetect iOS app target name for Expo pre-55 patch script phase
+
+- Updated dependencies [[`48358b2`](https://github.com/callstack/react-native-brownfield/commit/48358b2dcce578aa5052e66cc3454524da8c7992)]:
+  - @callstack/brownfield-cli@3.0.0
+
 ## 2.2.0
 
 ### Minor Changes